Info in Description (Routine Push)
Upgrades nearly done (just need to save data & load)
Working on decorations, store switching.
Added Firebase Crashlytics for the market crashes, have yet to get it to work.
Going to add a good amount of data for each player to the database, has to be clean.
Will be able to save/load store layouts w/ decorations and upgrades based on a slim dataset.